\name{GSelection-package}
\alias{GSelection-package}
\alias{GSelection}
\docType{package}
\title{
\packageTitle{GSelection}
}
\description{
\packageDescription{GSelection}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GSelection}
\packageIndices{GSelection}

}
\author{
\packageAuthor{GSelection}

Maintainer: \packageMaintainer{GSelection}
}
\references{
Guha Majumdar, S., Rai, A. and Mishra, D. C. (2019). Integrated framework for selection of additive and non-additive genetic markers for genomic selection. \emph{Journal of Computational Biology}. doi:10.1089/cmb.2019.0223
\cr Ravikumar, P., Lafferty, J., Liu, H. and Wasserman, L. (2009). Sparse additive models. \emph{Journal of the Royal Statistical Society: Series B (Statistical Methodology)}, 71(5), 1009-1030. doi:10.1111/j.1467-9868.2009.00718.x
\cr Yamada, M., Jitkrittum, W., Sigal, L., Xing, E. P. and Sugiyama, M. (2014). High-Dimensional Feature Selection by Feature-Wise Kernelized Lasso. \emph{Neural Computation}, 26(1):185-207. doi:10.1162/NECO_a_00537
}
